+++ src/CG.Reflection/Extensions/AppDomainExtensions.cs
@@ -15,6 +15,134 @@ public static partial class AppDomainExtensions
 
     /// <summary>
     /// This method searches among the assemblies loaded into the current 
+    /// app-domain for any matching concrete types that are assignable to 
+    /// the given type: <typeparamref name="T"/>.
+    /// </summary>
+    /// <typeparam name="T">The type to use for the search.</typeparam>
+    /// <param name="appDomain">The application domain to use for the operation.</param>
+    /// <param name="assemblyWhiteList">An optional white list of assembly
+    /// names - for narrowing the range of assemblies searched.</param>
+    /// <param name="assemblyBlackList">An optional white list of assembly
+    /// names - for narrowing the range of assemblies searched.</param>
+    /// <returns>A sequence of <see cref="Type"/> objects representing
+    /// zero or more matching concrete types.</returns>
+    public static IEnumerable<Type> FindConcreteTypes<T>(
+        this AppDomain appDomain,
+        string assemblyWhiteList = "",
+        string assemblyBlackList = "Microsoft*, System*, mscorlib, netstandard"
+        ) where T : class
+    {
+
+        // Validate the parameters before attempting to use them.
+        Guard.Instance().ThrowIfNull(appDomain, nameof(appDomain));
+
+        // Get the type we'll use in our search.
+        var searchType = typeof(T);
+
+        // Get the list of currently loaded assemblies.
+        var assemblies = AppDomain.CurrentDomain.GetAssemblies();
+
+        // Was a white list specified?
+        if (!string.IsNullOrEmpty(assemblyWhiteList))
+        {
+            // Look for assemblies in the white list we might need to load.
+            var toLoad = assemblyWhiteList.Split(',').Where(
+                x => assemblies.Any(y => !y.GetName().Name.IsMatch(x))
+                );
+
+            // Did we find any?
+            if (toLoad.Any())
+            {
+                // Loop and load any missing white listed assemblies.
+                toLoad.ForEach(x =>
+                {
+                    // Look for matching files.
+                    var files = Directory.GetFiles(
+                        AppDomain.CurrentDomain.BaseDirectory,
+                        x.EndsWith(".dll") ? x : $"{x}.dll"
+                        );
+
+                    // Loop through the files.
+                    foreach (var file in files)
+                    {
+                        try
+                        {
+                            Assembly.LoadFile(file);
+                        }
+                        catch (Exception)
+                        {
+                            // Don't care, just won't search this file.
+                        }
+                    }
+                });
+
+                // Get the list of currently loaded assemblies.
+                assemblies = AppDomain.CurrentDomain
+                    .GetAssemblies();
+
+                // Filter out anything that doesn't belong.
+                assemblies = assemblies.ApplyWhiteList(
+                    x => x.GetName().Name,
+                    assemblyWhiteList
+                    ).ToArray();
+            }
+
+            // At this point the assembly list should only contain assemblies 
+            //   that are already loaded in the app-domain, and have their 
+            //   name(s) on the white list.
+        }
+
+        // Was a black list specified?
+        if (!string.IsNullOrEmpty(assemblyBlackList))
+        {
+            // Split the black list into parts.
+            var blackParts = assemblyBlackList.Split(',');
+
+            // Filter out anything that doesn't belong.
+            assemblies = assemblies.ApplyBlackList(
+                x => x.GetName().Name,
+                assemblyBlackList
+                ).ToArray();
+
+            // At this point the assembly list should only contains those
+            //   assemblies that are already loaded in the app-domain and/or
+            //   white listed, and are not on the black list.
+        }
+
+        // Now we have a list of assemblies that contains juuuuuust the right items
+        //   and we can use that list to perform our next search.
+
+        var concreteTypes = new List<Type>();
+
+        // Loop and search.
+        foreach (var assembly in assemblies)
+        {
+            // Look for types that are public and non-abstract.
+            var types = assembly.GetTypes().Where(x =>
+                x.IsClass && !x.IsAbstract
+                );
+
+            // Loop through each matching type.
+            foreach (var type in types)
+            {
+                if (type.IsAssignableTo(searchType))
+                {
+                    concreteTypes.Add(type);
+                }
+            }
+        }
+
+        // Remove any duplicates.
+        var distinctList = concreteTypes.Distinct().ToList();
+
+        // Return the results.
+        return distinctList;
+    }
+
+    // *******************************************************************
+
+    /// <summary>
+    /// This method searches among the assemblies loaded into the current 
     /// app-domain for any public extension methods associated with the 
     /// specified type and signature.
     /// </summary>